Revert "refactor: migrate llm providers to ai.common.LLMBase (#719)"
This reverts commit b84a03a4fa599d51ab5417dfc18c4e2d5b148587.
