+++ nodes/src/nodes/accessibility_describe/IInstance.py
@@ -22,11 +22,11 @@
 # =============================================================================
 
 from .IGlobal import IGlobal
-from ai.common.llm_base import LLMBase
+from nodes.llm_base import IInstanceGenericLLM
 from rocketlib import AVI_ACTION
 
 
-class IInstance(LLMBase):
+class IInstance(IInstanceGenericLLM):
     """Instance handler for the Accessibility Scene Description node.
 
     Accepts image input and produces an accessibility-optimized text description
@@ -38,7 +38,6 @@ class IInstance(LLMBase):
     IGlobal: IGlobal
 
     def __init__(self):
-        """Initialize per-instance image buffer state."""
         super().__init__()
         self.image_data: bytearray | None = None
