+++ nodes/src/nodes/llm_base/IInstance.py
@@ -21,8 +21,30 @@
 # SOFTWARE.
 # =============================================================================
 
-from nodes.llm_base import IInstanceGenericLLM
+from rocketlib import IInstanceBase, invoke_function
+from ai.common.schema import Question, Answer
 
 
-class IInstance(IInstanceGenericLLM):
-    pass
+class IInstanceGenericLLM(IInstanceBase):
+    def _question(self, question: Question) -> Answer:
+        return self.IGlobal._chat.chat(question)
+
+    def writeQuestions(self, question: Question):
+        answer = self._question(question)
+        self.instance.writeAnswers(answer)
+
+    @invoke_function
+    def getContextLength(self, _param):
+        return self.IGlobal._chat.getTotalTokens()
+
+    @invoke_function
+    def getOutputLength(self, _param):
+        return self.IGlobal._chat.getOutputTokens()
+
+    @invoke_function
+    def getTokenCounter(self, _param):
+        return self.IGlobal._chat.getTokens
+
+    @invoke_function
+    def ask(self, param):
+        return self._question(param.question)
