+++ nodes/src/nodes/llm_ollama/IInstance.py
@@ -21,8 +21,8 @@
 # SOFTWARE.
 # =============================================================================
 
-from ai.common.llm_base import LLMBase
+from nodes.llm_base import IInstanceGenericLLM
 
 
-class IInstance(LLMBase):
+class IInstance(IInstanceGenericLLM):
     pass
